create RTDClass as separate module
