+++ examples/Temp_probes_RTD/Temp_probes_RTD.ino
@@ -16,8 +16,6 @@
 
 #include <Arduino_MachineControl.h>
 
-using namespace machinecontrol;
-
 // The value of the Rref resistor. Use 430.0 for PT100 and 4300.0 for PT1000
 #define RREF      400.0
 // The 'nominal' 0-degrees-C resistance of the sensor
